requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=py38h32f6830_1
av=8.0.2=py38he20a9df_1
backcall=0.2.0=pyh9f0ad1d_0
blas=2.16=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h1e0a361_1000
bzip2=1.0.8=h516909a_2
c-ares=1.16.1=h516909a_0
ca-certificates=2020.6.20=hecda079_0
cachetools=4.1.1=py_0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py38h924ce5b_2
cffi=1.14.0=py38hd463f26_0
chardet=3.0.4=py38h32f6830_1006
click=7.1.2=pyh9f0ad1d_0
cryptography=3.0=py38h766eaa4_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py38_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.3=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=he06d7ca_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.65.0=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
google-auth=1.19.2=py_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.14=h23475e2_0
grpcio=1.30.0=py38h2c89da0_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_100
icu=64.2=he1b5a44_1
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
importlib-metadata=1.7.0=py38h32f6830_0
intel-openmp=2020.1=217
ipdb=0.13.0=pyh9f0ad1d_0
ipython=7.16.1=py38h23f93f0_0
ipython_genutils=0.2.0=py_1
jasper=1.900.1=h07fcdf6_1006
jedi=0.17.2=py38h32f6830_0
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py38hfd86e86_0
lame=3.100=h14c3975_1001
lcms2=2.11=hbd6801e_0
ld_impl_linux-64=2.34=h53a641e_7
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=9.0.1=default_hde54327_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1006
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libllvm9=9.0.1=he513fc3_1
libopenblas=0.3.10=h5a2b251_0
libopencv=4.4.0=py38_2
libpng=1.6.37=hed695b0_1
libprotobuf=3.12.3=h8b12597_2
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc7e4089_6
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.14=h7b6447c_0
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he1b5a44_1
markdown=3.2.2=py_0
matplotlib=3.3.0=1
matplotlib-base=3.3.0=py38h91b0d89_1
mkl=2020.1=217
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.5=pypi_0
ninja=1.10.0=hc9558a2_0
nspr=4.27=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.19.1=py38h8854b6b_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
opencv=4.4.0=py38_2
openh264=2.1.1=h8b12597_0
openssl=1.1.1h=h516909a_0
pandas=1.0.5=py38h0573a6f_0
parso=0.7.1=pyh9f0ad1d_0
patsy=0.5.1=py_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=py38h32f6830_1
pickleshare=0.7.5=py38h32f6830_1001
pillow=7.2.0=py38h9776b28_1
pip=20.1.1=py38_1
pixman=0.38.0=h516909a_1003
prompt-toolkit=3.0.5=py_1
protobuf=3.12.3=py38h950e882_0
ptyprocess=0.6.0=py_1001
py-opencv=4.4.0=py38h23f93f0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.6.1=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.5=h425cb1d_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.4.0=py3.8_cuda10.0.130_cudnn7.6.3_0
pytz=2020.1=py_0
qt=5.12.5=hd8c4c69_1
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
scipy=1.5.2=py38h8c5af15_0
seaborn=0.11.0=0
seaborn-base=0.11.0=py_0
setuptools=49.2.0=py38_0
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
statsmodels=0.12.0=py38hab2c0dc_1
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
tk=8.6.10=hbc83047_0
torchvision=0.5.0=py38_cu100
tornado=6.0.4=py38h7b6447c_1
traitlets=4.3.3=py38h32f6830_1
urllib3=1.25.10=py_0
wcwidth=0.2.5=pyh9f0ad1d_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py38_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h6597ccf_1
